feat: 수정 후 스와이프 닫기
